clean up, added options, more tests
